requirements.txt
aiohttp
asyncssh
asyncstdlib
argon2-cffi
beautifulsoup4==4.12.3
black
bokeh
numpy
channels
django>=5.0.0
django-cms
djangocms-attributes-field
django_compressor
django_compressor_toolkit
django-configurations>=2.4
django-cors-headers
django-easy-logging
django-extensions
django-filer
django-filter
django-memoize
django-minio-storage
django-pipeline
django-prometheus
django-recaptcha==4.0.0
django-regex-field
djangorestframework
django-stubs
django-stubs-ext
django-tables2
django-treebeard
django-types
django_user_agents
django-otp
django-two-factor-auth
django-mfa2
easy_thumbnails
exceptionite
flake8-django
geopy
git+https://github.com/holdenk/static-thumbnails.git
git+https://github.com/jazzband/django-cookie-consent.git#egg=django-cookie-consent
git+https://github.com/totallylegitco/django-encrypted-model-fields.git@87dfff140b3ad256404a913019e90f2c245e57f8
git+https://github.com/totallylegitco/llm-result-utils.git@7e2feb036bb40f6c93214b1b1282684a9b33ccda
icd10-cm
loguru
markdown_strings
metapub==0.5.12
mypy
nest_asyncio
packaging>=20.0
pandoc
paramiko
Pillow
psycopg2
psycopg-pool
pyarrow
pymupdf
pyOpenSSL
PyPDF2
pyright
pytesseract # We either need easyocr + opencv-python OR pytesseract + tesseract (so alt is easyocr>=1.7.0 opencv-python)
python-Levenshtein
pytz
pyyaml
RapidFuzz==2.15.1 # See https://github.com/mindee/doctr/issues/1186
ray==2.38.0
ray[client]==2.38.0
requests
sentry-sdk[django]
setuptools
sorl-thumbnail
sqlalchemy_mate==2.0.0.0 # uszipcode issue see https://github.com/MacHu-GWU/uszipcode-project/issues/79
stopit
stripe>11.0
totallylegit-django-rest-framework-braces # forked from django-rest-framework-braces
types-paramiko
types-requests
typing-extensions
urllib3
uszipcode
uvicorn[standard]
werkzeug
